+++ node/src/Commands.ts
@@ -2019,5 +2019,9 @@ export function createZIncrBy(
     increment: number,
     member: string,
 ): command_request.Command {
-    return createCommand(RequestType.ZIncrBy, [key, increment.toString(), member]);
+    return createCommand(RequestType.ZIncrBy, [
+        key,
+        increment.toString(),
+        member,
+    ]);
 }